added removing nodes from graph
added an array indicating if a node is active or not as well as methods
to get all the active nodes and check if a node is valid, as well as
methods to make nodes inactive.

also made other methods play well with active nodes.
get_predecessors and get_neighbours check if predecessor or neighbour
nodes are active before adding them to the arrays.