implemented scc algorithn using pthreads
created scc_pthreads header file that contains implementation for the
scc graph coloring algorithm in parallel using pthreads. the algorithm
runs correctly and with good performance, but the documentation is
unfinished and outdated. will attempt to fix this tomorrow.

also changed main.c to print only the number of sccs returned using each
implementation and print the vertex ids only if its different between
implementations. more importantly, now main.c measures the time elapsed
during the execution of the implementations in milliseconds and prints
it to the console.

Makefile was updated to include the new files

in scc.c minor change in how n_active_verts is updated in the final
loop. it probably didnt affect performance too much but after noticing
it was bothering me so i changed it.